refactored file positions and fixed bugs
