+++ backend/models/posting.js
@@ -9,12 +9,12 @@ var PostingSchema = new Schema({
   anonName: String,
   createdAt: { type: Date, default: Date.now },
   replies: [{
-  				 rid: Number, 
-  				 content : String, 
-  				 user : { type: Number, ref: 'Account' },
-  				 anonName: String,
-  				 createdAt: { type: Date, default: Date.now }
-  				 }]
+    rid: Number, 
+    content : String, 
+    user : { type: Number, ref: 'Account' },
+    anonName: String,
+    createdAt: { type: Date, default: Date.now }
+ }]
 });
 
 module.exports = mongoose.model('Posting', PostingSchema);
