+++ routes/feed.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var router = express.Router();
-var logger = require('../backend/logger.js');
+var logger = require('../backend/services/logger.js');
 
 
 /* GET posting feed 
@@ -9,4 +9,6 @@ query: limit
 
 router.get('/', function(req, res, next) {
 
-}
+});
+
+module.exports = router;
