+++ routes/user.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var router = express.Router();
-var logger = require('../backend/logger.js');
+var logger = require('../backend/services/logger.js');
 
 /* GET info and posting of a user 
 example: {
@@ -13,7 +13,7 @@ example: {
 */
 router.get('/:id', function(req, res, next) {
 
-}
+});
 
 
 module.exports = router;
